classes moved to platform test classes
