+++ platform-sdk/swirlds-common/src/testFixtures/java/com/swirlds/common/test/fixtures/threading/ConcurrentTesting.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.common.test.threading;
+package com.swirlds.common.test.fixtures.threading;
 
 import static com.swirlds.common.test.fixtures.AssertionUtils.assertEventuallyTrue;
