+++ platform-sdk/swirlds-common/src/testFixtures/java/module-info.java
@@ -1,8 +1,10 @@
-module com.swirlds.common.test.fixtures {
+open module com.swirlds.common.test.fixtures {
     exports com.swirlds.common.test.fixtures;
     exports com.swirlds.common.test.fixtures.crypto;
     exports com.swirlds.common.test.fixtures.io;
     exports com.swirlds.common.test.fixtures.merkle.util;
+    exports com.swirlds.common.test.fixtures.threading;
+    exports com.swirlds.common.test.fixtures.stream;
 
     requires com.swirlds.common;
     requires org.junit.jupiter.api;
