+++ platform-sdk/swirlds-platform-core/build.gradle.kts
@@ -52,6 +52,7 @@ dependencies {
     testImplementation(testLibs.bundles.mocking)
     testImplementation(testLibs.bundles.utils)
     testImplementation(project(":swirlds-config-impl"))
+    testImplementation(project(":swirlds-unit-tests:core:swirlds-platform-test"))
     testImplementation(testFixtures(project(":swirlds-base")))
     testImplementation(testFixtures(project(":swirlds-common")))
     testImplementation(testFixtures(project(":swirlds-config-api")))
