+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/state/SwirldStateManagerUtilsTests.java
@@ -23,8 +23,8 @@ import static org.mockito.Mockito.when;
 
 import com.swirlds.common.system.BasicSoftwareVersion;
 import com.swirlds.common.system.SwirldState;
-import com.swirlds.common.test.state.DummySwirldState;
 import com.swirlds.platform.metrics.SwirldStateMetrics;
+import com.swirlds.platform.test.state.DummySwirldState;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
