+++ platform-sdk/swirlds-unit-tests/common/swirlds-common-test/src/main/java/module-info.java
@@ -1,12 +1,10 @@
 open module com.swirlds.common.test {
-    exports com.swirlds.common.test.state;
     exports com.swirlds.common.test.merkle.util;
     exports com.swirlds.common.test.merkle.dummy;
     exports com.swirlds.common.test.dummy;
     exports com.swirlds.common.test.benchmark;
     exports com.swirlds.common.test.set;
     exports com.swirlds.common.test.map;
-    exports com.swirlds.common.test.threading;
     exports com.swirlds.common.test.fcqueue;
 
     requires com.swirlds.base;
