+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/sync/SyncTests.java
@@ -29,8 +29,8 @@ import static org.mockito.Mockito.when;
 import com.swirlds.common.constructable.ConstructableRegistry;
 import com.swirlds.common.constructable.ConstructableRegistryException;
 import com.swirlds.common.system.NodeId;
-import com.swirlds.common.test.threading.ReplaceSyncPhaseParallelExecutor;
-import com.swirlds.common.test.threading.SyncPhaseParallelExecutor;
+import com.swirlds.common.test.fixtures.threading.ReplaceSyncPhaseParallelExecutor;
+import com.swirlds.common.test.fixtures.threading.SyncPhaseParallelExecutor;
 import com.swirlds.common.threading.pool.CachedPoolParallelExecutor;
 import com.swirlds.common.threading.pool.ParallelExecutionException;
 import com.swirlds.common.threading.pool.ParallelExecutor;
